screen class updates 5/7
made the screen class adapatable for dealer as well. I would like to cut down on the repetition of creating a new screen instance, but for now it works. Next up is creating a main class in which I will move over the main code into a game class that will allow for multiple instances of the game that can continue until the user no longer wants to.